GOCBC-725: Fixed inconsistent use of integer bit-specifiers.
Reviewed-on: http://review.couchbase.org/120594
Tested-by: Brett Lawson <brett19@gmail.com>

+++ bucket_ping.go
@@ -39,7 +39,7 @@ type jsonEndpointPingReport struct {
 }
 
 type jsonPingReport struct {
-	Version  int                                 `json:"version"`
+	Version  uint16                              `json:"version"`
 	SDK      string                              `json:"sdk,omitempty"`
 	ID       string                              `json:"id,omitempty"`
 	Services map[string][]jsonEndpointPingReport `json:"services,omitempty"`
